Fixed certificate serial overflow
Certificate serial numbers and CRL serial numbers are now stored in
SQLite as a string. When retrieved from the database, they are cast
as a Python int so we can easily increment the number. This solves
the problem where a certificate with a large serial number is imported
into the CA.